requirements.txt
# Market2Agent — All Dependencies

# Core Framework
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
pydantic>=2.5.0,<3.0.0

# Async HTTP
httpx>=0.27.0,<1.0.0

# DNS & Domain Analysis
dnspython>=2.5.0,<3.0.0
python-whois>=0.9.0,<1.0.0

# Database
neo4j>=5.17.0,<6.0.0

# Caching & Rate Limiting
redis>=5.0.0,<6.0.0

# Background Workers
arq>=0.25.0,<1.0.0

# Auth
PyJWT>=2.8.0,<3.0.0

# Payments
stripe>=8.0.0,<10.0.0

# Logging
structlog>=24.1.0,<25.0.0

# Web Scraping
beautifulsoup4>=4.12.0,<5.0.0
lxml>=5.1.0,<6.0.0

# Config
python-dotenv>=1.0.0,<2.0.0

# Form/File Uploads
python-multipart>=0.0.9,<1.0.0
